Cambios minimos

+++ mvc-go/dto/dto_order.go
@@ -2,7 +2,7 @@ package dto
 
 type OrderDto struct {
 	IdOrder     int            `json:"id_order"`
-	MontoFinal  int            `json:"monto_final"`
+	MontoFinal  float32        `json:"monto_final"`
 	Fecha       string         `json:"fecha"`
 	Estado      string         `json:"estado"`
 	Usuario     UserDto        `json:"Usuario"`
