+++ mvc-go/dto/dto_order_detail.go
@@ -2,9 +2,9 @@ package dto
 
 type OrderDetailDto struct {
 	IdOrderDeatail int        `json:"id_order_detail"`
-	PrecioUnitario int        `json:"precio_unitario"`
+	PrecioUnitario float32    `json:"precio_unitario"`
 	Cantidad       int        `json:"cantidad"`
-	Total          int        `json:"total"`
+	Total          float32    `json:"total"`
 	Detalle        string     `json:"detalle"`
 	Producto       ProductDto `json:"Product"`
 }
